[audio hal]: Modify the bugs which EXOPlayer can't play files when the format              of audio is ac3/eac3.
1. This is not a correct in our codes of audio hal to report the support formats.
The key of AUDIO_PARAMETER_STREAM_SUP_FORMATS is query the formats can be
send to audiotrack, for example, if we return AUDIO_FORMAT_AC3, the exoplayer
will send raw ac3 datas to audiotrack, and this format we can't decode.
This will lead to create a audiotrack fail,and play this file fail.
2. Add a interface to query which formats that hdmi can bitstream.

Change-Id: I307e7f0fa563b3114f832b8ca45d60efd587c43d
Signed-off-by: hehua <hh@rock-chips.com>